Extract email AI stream handling
